backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
joblib==1.3.1
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0